pallets/camp-assets/src/tests.rs
use crate::{mock::*, Error};
use frame_support::{assert_noop, assert_ok};

#[test]
fn it_works_for_creating_a_token() {
	new_test_ext().execute_with(|| {
		// Dispatch a signed extrinsic.
		assert_ok!(TemplateModule::create_token(
			Origin::signed(1),
			b"batman".to_vec(),
			b"BAT".to_vec(),
			18,
			10_000
		));
		// Read pallet storage and assert an expected result.
		assert_eq!(Assets::metadata(0), Some(42));
	});
}

#[test]
fn correct_error_for_none_value() {
	new_test_ext().execute_with(|| {
		// Ensure the expected error is thrown when no value is present.
		// assert_noop!(TemplateModule::cause_error(Origin::signed(1)), Error::<Test>::NoneValue);
	});
}
